+++ src/Symfony/Component/DomCrawler/Crawler.php
@@ -564,10 +564,8 @@ class Crawler implements \Countable, \IteratorAggregate
      *
      * @param bool $normalizeWhitespace Whether whitespaces should be trimmed and normalized to single spaces
      */
-    public function innerText(/* bool $normalizeWhitespace = true */): string
+    public function innerText(bool $normalizeWhitespace = true): string
     {
-        $normalizeWhitespace = 1 <= \func_num_args() ? func_get_arg(0) : true;
-
         foreach ($this->getNode(0)->childNodes as $childNode) {
             if (\XML_TEXT_NODE !== $childNode->nodeType) {
                 continue;